remove `rust_decimal` After extensive testing, it has been concluded that this type does not yet provide any useful benefits. Furthermore, it offers worse performance and uses more memory.

+++ src/aliases.rs
@@ -1,7 +1,5 @@
-use rust_decimal::Decimal;
-
 /// A fixed-precision decimal number that represents temperature in *Celsius*.
-pub type Temperature = Decimal;
+pub type Temperature = f32;
 
 /// Relative humidity represented as percentage (*%*).
 pub type Humidity = u8;
@@ -10,7 +8,7 @@ pub type Humidity = u8;
 pub type AirPressure = u16;
 
 /// A fixed-precision decimal number that represents voltage (*V*).
-pub type BatteryVoltage = Decimal;
+pub type BatteryVoltage = f32;
 
 /// Signal strength represented in decibel-milliwatts (*dBm*).
 pub type Rssi = i8;
